refactor to handle dynamic seeds and proportions
Users of the cline-cutter can now specify an arbritrary number of seeds and proportions to test in the pipeline command line. The pipeline will then keep track of this information for each combination of seed, proportion, and downsampling regime. So, if the user specifies X seeds, Y proportions, and Z downsampling regimes, the pipeline will run cline modeling X * Y * Z times.